pierrick reviews and add a new component for editing mapping
